+++ src/frontend/src/pages/Discover/ModalsWrapper.tsx
@@ -1,6 +1,8 @@
 import EditFiltersModal from "components/EditFiltersModal";
 import EditTripsModal from "components/EditTripsModal";
+import EditUserProfileWizard from "components/EditWizard/EditUserProfileWizard";
 import SearchModal from "components/SearchModal";
+import EditTripsWizard from "src/components/EditWizard/EditTripsWizard";
 import NeighborhoodDetail from "../NeighborhoodDetail";
 
 const ModalsWrapper = ({ isMobile }: { isMobile?: boolean }) => (
@@ -9,6 +11,8 @@ const ModalsWrapper = ({ isMobile }: { isMobile?: boolean }) => (
         <SearchModal isMobile={isMobile} />
         <EditFiltersModal isMobile={isMobile} />
         <EditTripsModal isMobile={isMobile} />
+        <EditUserProfileWizard />
+        <EditTripsWizard />
     </>
 );