+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add address autocomplete and handle invalid destinations [#742](https://github.com/azavea/echo-locator/pull/742)
 - Add edit trip modal [#744](https://github.com/azavea/echo-locator/pull/744)
 - Add neighborhood search [#743](https://github.com/azavea/echo-locator/pull/743)
+- Enable profile/trip wizards edit mode [#745](https://github.com/azavea/echo-locator/pull/745)
 
 ### Changed